chore(android): add missing result.success call
The flutter side of the MethodChannel returns a Future, thus most developers
would add an await. But since the result.success() method is never called
the Future will never resolve and it will wait forever.

To make the implementation match the interface lets just call
result.success(null) when done to indicate that the method finished.